Add plot to fold detail view using react-vis. This diff only includes functionality, not styling.
Summary:
        - Add react-vis
        - Add xyplot with bubble marks
        - Add axis labels
        - Move search into separate helper file
        - Connect data, including search and filters
        - Add tooltips
        - Add zoom

Test Plan:
Run react app. Click 'view all' on a pulsar. You should see a
graph displayed. Hover on a bubble for a tooltip. Use the Search/
filters to change the plot data. Clicking and dragging on the plot
should zoom in. Clicking a blank spot on the chart should reset it.

Reviewers: sosl, llakerink

Reviewed By: sosl

Differential Revision: https://phab.adacs.org.au/D345