Adjust the DB schema, restore pulsar detail view
Summary:
- DB adjusted to match https://dbdiagram.io/d/5fc70be63a78976d7b7e284f (version 11:47 am, 03/12/20)
- Tweak ingest to match the new DB schema
- Tweak graphql schema to match the new layout
- Mostly restore the pulsar detail view

Test Plan:
This code is still a bit of WIP but should function. Tested on the data in test-data. You should be able to ingest the data, then go to e.g. http://localhost:8000/J1424-5556 and see the observation in a table, with DM_ SN_ and RM_ meerpipe missing.

```
poetry run pytest meertime/test_graphql.py
```
should succeed

Reviewers: ajameson

Reviewed By: ajameson

Differential Revision: https://phab.adacs.org.au/D365