Fix a bug in psrdb.py: setting url
Summary: If custom url was provided, it was provided as a list which didn't work properly. Fixed by changing default to a list, and extracting the first element of the list in rest of the code.

Test Plan:
```
poetry run pytest psrdb.py -u http://your_url/graphql/ -t "$token" pulsars list
```

Reviewers: ajameson

Reviewed By: ajameson

Differential Revision: https://phab.adacs.org.au/D377