Add pipelineimages mutations, table in CLI, storage tests
Summary: pipelineimages is a bit special as we're uploading a file to ImageField which is not really supported directly by GraphQL. We do it by encoding the contents in base64 to ensure all characters are suitable for sending via requests and send them as a string. We then decode the contents back from base64 and wrap them in a ContentFile which can be directly saved in ImageField. We use a custom storage class to overwrite images if a newer version is uploaded.

Test Plan:
Note that model_bakery was added as a dev dependency so you'll need to install it with poetry. It's a module which makes creating model instances, along with dependendt objects very easy. See docs or dataportal/test_storage.py for an example usage

```
pytest dataportal/test_storage.py
```
should pass

```
python psrdb.py -u $url -t $token pipelineimages list
python psrdb.py -u $url -t $token pipelineimages create test 1 99 your_image.png
```
should both work. Note that you may need to tweak the create command to match an existing processing id (that's the "1" in the command)

Reviewers: ajameson

Reviewed By: ajameson

Differential Revision: https://phab.adacs.org.au/D445