Update frontend to new version.
Summary: This is a straight swap of the react frontend from the current master branch. The react app hasn't been kept up to date with master and was in an old incomplete state. This fixes that issue. All code has been previously reviewed.

Test Plan: Unfortunately there is currently no way to test this as the backend doesn't work. You can run 'npm test' which should have passing tests.

Reviewers: ajameson, ssaleheen

Reviewed By: ajameson

Differential Revision: https://phab.adacs.org.au/D631