Change how images are collected by the web cache.
Summary: Images are stored differently than originally thought. This diff changes the urls that the frontend consumes so that they point to the correct location. This should fix the 404 urls.

Test Plan: Run the backend and frontend. Go to Frontend in a browswer. Click > Fold mode > View Last. The images should point to the correct path (but they'll still error as the images aren't stored locally).

Reviewers: ssaleheen, ajameson

Reviewed By: ssaleheen, ajameson

Differential Revision: https://phab.adacs.org.au/D657